close #5 create comp & gaibuhenokakidashi
