+++ react-ts-app/src/index.tsx
@@ -1,46 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import reportWebVitals from './reportWebVitals';
+import App from './App';
 
-function formatName(user: User){
-  return `${user.firstName} ${user.lastName}`;
-}
-
-interface User {
-  firstName: string;
-  lastName: string;
-  avatarUrl: string;
-}
-
-const user: User = {
-  firstName: "toma",
-  lastName: "yu",
-  avatarUrl: "https://placehold.jp/150x150.png"
-}
-
-function getGreeting() {
-  const greetingName = user? formatName(user): "stranger";
-  return <h1>hello, {greetingName}!</h1>
-}
-
-const elementGreeting = getGreeting();
-const elementImg = <img src={user.avatarUrl} alt="test" />
-const elementCreateElement = React.createElement(
-  "h1",
-  {className: "greeting"},
-  "こんにちは"
-)
-
-const element = (
-  <div>
-    <div id="1">{elementGreeting}</div>
-    <div id="2">{elementImg}</div>
-    <div id="3">{elementCreateElement}</div>
-  </div>
-);
 
 ReactDOM.render(
-  element,
+  <App/>,
   document.getElementById("root")
 )
